vote post

+++ database/migrations/2015_03_26_140533_create_posts_table.php
@@ -22,7 +22,7 @@ class CreatePostsTable extends Migration {
             $table->string('url');
             $table->string('query_url');
             $table->string('approval', 1)->default('T');
-            $table->integer('votes');
+            $table->decimal('votes', 3, 2);
             $table->integer('clicks');
             $table->timestamps();
         });
